Add custom control panel class
Currently changing program functionality requires manual programmatic change.

We change this to a centralized unified control panel class that allows us to
set the behaviour that we want from our application and that will determine the
program flow accordingly.

For example, we change IS_SIMULATION_ACTIVE to true if we want to add a window event
listener that will trigger a test sequence upon page loading, and we set this to false
if we want to directly access the puppet without any simulation.